style & some essential fixes to initialize the service
